--- src/App.css
@@ -1,145 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 10vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.app-title {
-  font-size: 50px;
-  margin-bottom: 15px;
-  color: white;
-}
-
-.app-container {
-  color: white;
-  position: absolute;
-  -webkit-transform: translate(0%, -50%);
-  transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  left: 50%;
-  top: 50%;
-  width: 100%;
-  text-align: center;
-  background-image: url(Assets/food.jpg);
-  background-size: cover;
-  background-repeat: no-repeat;
-  box-shadow:inset 0 0 0 2000px rgba(0, 0, 0, 0.8);
-}
-
-.btns {
-  margin-top: 25%;
-  width: 100%;
-  height: 100vh;  
-}
-
-.btn-outlined {
-  display: flex;
-  justify-content: center;
-  margin: 2%;
-}
-
-button {
-  color: white;
-  margin: 20px 15px 0 0;
-  font-size: 14px;
-  text-decoration: none;
-  padding: 10px;
-  background: transparent;
-  border: 1px solid white;
-  border-radius: 5px;
-  min-width: 100px;
-}
-
-.links {
-  color: white;
-  text-decoration: none;
-}
-
-.icons {
-  margin-right: 15px;
-  font-size: 12px;
-  color: white;
-  vertical-align: sub;
-}
-
-button:hover {
-  border: 1px solid transparent;
-  background-color: blue;
-  color: white;
-}
-
-button:focus {
-  box-shadow: unset;
-  border-color: white;
-  background-color: unset;
-  color: white;
-}
-
-.login-page {
-  height: 100vh;
-  display: block;
-  text-align: center;
-}
-
-.login-bg-img {
-  background-image: url(Assets/food.jpg);
-  background-Repeat: no-repeat;
-  background-Color:grey;
-  background-size: cover;
-  background-position: center;
-}
-
-.login-paper {
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.login-avt {
-  margin: 5%;
-}
-
-.login-form {
-  width: 100%;
-  margin: 5%;
-  padding: 4%;
-}
-
-.sbmit-signup {
-  margin-top: 2%;
-  margin-bottom: 2%;
-}
