+++ src/Components/App.js
@@ -1,20 +1,17 @@
 import React from 'react';
-import '../App.css';
 import { Route, Switch } from 'react-router-dom';
-import Main from './LandingPage';
-import NavBar from '../Containers/Nav';
-import AllMeals from '../Containers/Recipes';
+import AllMeals from '../Containers/AllRecipes';
 import MealDetails from '../Containers/ItemDetails';
+import NavBar from '../Containers/Nav';
 import AllCategories from '../Containers/Explore';
 
 const App = () => (
-  <div className="App">
+  <div>
     <NavBar />
     <Switch>
-      <Route path="/" component={Main} exact />
-      <Route path="/explore" component={AllCategories} exact />
+      <Route path="/" component={AllCategories} exact />
       <Route path="/category/:category" component={AllMeals} />
-      <Route path="/recipe/:id" component={MealDetails} />
+      <Route path="/meal/:id" component={MealDetails} />
     </Switch>
   </div>
 );
