+++ src/Components/Recipes.js
@@ -2,12 +2,12 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReciperImgage from './RecipeImage';
+import Image from './RecipeImage';
 
 const Meal = ({
   src, name,
 }) => (
-  <ReciperImgage className="meal-preview shadow" src={src} name={name} alt={name} />
+  <Image className="meal-preview shadow" src={src} name={name} alt={name} />
 );
 
 Meal.propTypes = {
