+++ src/Settings/Index.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-plusplus */
 export const getProducts = state => state.products;
 export const getProductsPending = state => state.pending;
 export const getProductsError = state => state.error;
@@ -13,6 +12,7 @@ export const categoryImage = category => category.strCategoryThumb;
 
 export const getIngredient = meal => {
   const result = [];
+  // eslint-disable-next-line no-plusplus
   for (let i = 1; i <= 20; i++) {
     if (meal[`strIngredient${i}`] === '') {
       break;
